+++ storage/archive_storage.go
@@ -1,4 +1,4 @@
-package minio
+package storage
 
 import (
 	"context"
